feat: Add nudge creation job to cronjob Restructure codes to enforce order of two cronjobs
